requirements.txt
asgiref==3.8.1
attrs==23.2.0
Automat==22.10.0
beautifulsoup4==4.12.3
black==23.12.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
cryptography==42.0.8
cssselect==1.2.0
debugpy==1.8.0
defusedxml==0.7.1
Django==5.0.6
docutils @ file:///private/tmp/docutils-20231211-36728-9mhbst/docutils-0.20.1
filelock==3.14.0
greenlet==3.0.3
h11==0.14.0
hyperlink==21.0.0
idna==3.6
imageio==2.33.1
incremental==22.10.0
itemadapter==0.9.0
itemloaders==1.3.1
jmespath==1.0.1
lazy_loader==0.3
lxml==5.1.0
msgpack==1.0.7
mypy-extensions==1.0.0
neovim==0.3.1
networkx==3.2.1
numpy==1.26.3
opencv-python==4.9.0.80
outcome==1.3.0.post0
packaging==23.2
parsel==1.9.1
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.1.0
Protego==0.3.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyDispatcher==2.0.7
pynvim==0.5.0
pyOpenSSL==24.1.0
PySocks==1.7.1
pytesseract==0.3.10
python-dotenv==1.0.1
PyYAML @ file:///private/tmp/pyyaml-20231018-5610-1rdh755/PyYAML-6.0.1
queuelib==1.7.0
requests==2.31.0
requests-file==2.1.0
scikit-image==0.22.0
scipy==1.11.4
Scrapy==2.11.2
selenium==4.16.0
service-identity==24.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.0
tifffile==2023.12.9
tldextract==5.1.2
trio==0.24.0
trio-websocket==0.11.1
Twisted==24.3.0
typing_extensions==4.12.2
urllib3==2.1.0
w3lib==2.2.0
webdriver-manager==4.0.2
wsproto==1.2.0
zope.interface==6.4.post2
